+++ src/Opulence/Databases/Migrations/MigrationResolutionException.php
@@ -1,0 +1,21 @@
+<?php
+
+/**
+ * Opulence
+ *
+ * @link      https://www.opulencephp.com
+ * @copyright Copyright (C) 2017 David Young
+ * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
+ */
+
+namespace Opulence\Databases\Migrations;
+
+use Exception;
+
+/**
+ * Defines a migration resolution exception
+ */
+class MigrationResolutionException extends Exception
+{
+    // Don't do anything
+}